[level 1] Title: 콜라츠 추측, Time: 0.02 ms, Memory: 64.5 MB -BaekjoonHub

+++ 프로그래머스/1/12943. 콜라츠 추측/콜라츠 추측.java
@@ -1,5 +1,5 @@
 class Solution {
-    public int solution(int num) {
+    public int solution(long num) {
         if(num == 1){
             return 0;
         }
